requirements.txt
Django==1.11
djangorestframework==3.11.0
numpy==1.19.1
pandas==1.1.0
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
